pyproject.toml
[project]
name = "Spellcheck"
version = "0.1.0"
description = "Spellcheck: A tool that offers mastery over the world of writing"
authors = [
    {name = "Kanishk Pachauri", email = "itskanishkp.py@gmail.com"},
]
dependencies = [
    "fastapi>=0.110.0",
    "sqlalchemy>=2.0.27",
    "pydantic-settings>=2.2.1",
    "uvicorn>=0.27.1",
    "spello>=1.3.0",
    "requests>=2.31.0",
    "rich>=13.7.0",
    "pandas>=2.2.1",
    "python-dotenv>=1.0.1",
    "black>=24.2.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
start = "uvicorn spellcheck.app:app --host 0.0.0.0 --port 8080"
dev = "uvicorn spellcheck.app:app --host 0.0.0.0 --port 8080 --reload"
download_model = "python3 language_models/download_models.py"
format = "black spellcheck"